[deploy] update managed clusters in flux clusters
the ClusterDeployment resource wants its cluster's baseDomain and
cloud region explicitly defined. I don't know why it requires this given
that these are defined in the ClusterDeployment's ref'ed InstallConfig,
but this commit updates the deploy script so that it automatically
updates these whenever they are changed in config.yaml. Quite annoying
given the amount of duplication, but it is what it is.

Signed-off-by: Carlos Nunez <cnunez@redhat.com>